feat(chat-service): 채팅 이미지 저장 메서드 이관 및 예외 처리 추가 #178
